Extract scan module
