+++ src/scan/mod.rs
@@ -1,0 +1,10 @@
+use std::path::*;
+
+pub mod path_suffix;
+
+#[derive(Debug, PartialEq)]
+pub struct CommandVersion {
+    pub command: String,
+    pub version: String,
+    pub path: PathBuf,
+}
