+++ src/scan_cmd.rs
@@ -9,29 +9,8 @@ use regex::Regex;
 use std::process;
 use def_file;
 use shim;
-
-lazy_static! {
-    static ref COMMAND_VERSION_REGEX: Regex =
-        Regex::new(r"^(?P<command>.+[^\d.-])-?(?P<version>\d+(?:\.\d+)*)$").unwrap();
-}
-
-#[derive(Debug, PartialEq)]
-struct CommandVersion {
-    command: String,
-    version: String,
-    path: PathBuf,
-}
-
-fn parse_command_version(bin: PathBuf) -> Option<CommandVersion> {
-    let name = String::from(bin.file_name().unwrap().to_str().unwrap());
-
-    COMMAND_VERSION_REGEX.captures(&name)
-        .map(|captures| CommandVersion {
-            command: String::from(captures.name("command").unwrap().as_str()),
-            version: String::from(captures.name("version").unwrap().as_str()),
-            path: bin
-        })
-}
+use scan;
+use scan::CommandVersion;
 
 fn prompt_versions(versions: &Vec<CommandVersion>) -> Vec<usize> {
     let items: Vec<_> = versions.iter()
@@ -56,17 +35,14 @@ fn prompt_versions(versions: &Vec<CommandVersion>) -> Vec<usize> {
 }
 
 pub fn run(command: &str) {
-    let shim_dir = config::shim_dir();
-    let path = env::var("PATH").expect("env var PATH is not defined");
-    let versions: Vec<_> = env::split_paths(&path)
-        .filter(|p| p != &shim_dir)
-        .flat_map(|p| fs::read_dir(p).unwrap())
-        .map(|bin| bin.unwrap().path())
-        .flat_map(|bin| parse_command_version(bin))
-        .filter(|c| c.command == command)
-        .collect();
+    let scans = vec![
+        scan::path_suffix::scan(command)
+    ];
+    let is_empty = scans.iter()
+        .all(|v| v.is_empty());
+    let versions: Vec<_> = scans.into_iter().flat_map(|x| x).collect();
 
-    if versions.is_empty() {
+    if is_empty {
         println!("Sorry, could not find any versions of {}", command);
         process::exit(1);
     } else {
